% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thunder.R
\name{get_props}
\alias{get_props}
\title{Get cell type proportions from THUNDER fit}
\usage{
get_props(thunder_fit)
}
\arguments{
\item{thunder_fit}{Output of `thunder`}
}
\description{
Get cell type proportions from THUNDER fit
}
