% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodograph.R
\name{hodograph}
\alias{hodograph}
\title{Plot hodograph}
\usage{
hodograph(
  u,
  v,
  hght,
  max_hght = 16000,
  max_speed = 25,
  lab_hghts = c(0, 1, 3, 6, 9, 12),
  ...
)
}
\arguments{
\item{u}{u-wind wind vector components (m/s)}

\item{v}{v-wind wind vector componente (m/s)}

\item{hght}{vector of altitudes (m)}

\item{max_hght}{max. altitude to be considered. 8 km used by default}

\item{max_speed}{max. isoline (circle) to be drawn for wind speed; 30 m/s used as default}

\item{lab_hghts}{label of numeric vector heights in km to be signed on hodograph; default: 1, 3, 6; NULL for skipping labels}

\item{...}{other graphical parameters to be used with plot() function}

\item{by_km}{logical, whether to draw changes in wind profile by 1 km interval (default); alternatively every single entry will be drawn}
}
\description{
Plot hodograph to show changes of wind speed and direction in vertical profile
}
\examples{
data("sounding_wien")
attach(sounding_wien)
   # changing wind speed and direction to U and V wind components
   # also changing units from knots to m/s
   u = round(-SKNT * 0.514444 * sin(DRCT * pi/180), 2)
   v = round(-SKNT * 0.514444 * cos(DRCT * pi/180), 2)
   # finally plot the hodograph:
   hodograph(u, v, HGHT)

}
