% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_nmf_fit.R
\name{initial_nmf_fit}
\alias{initial_nmf_fit}
\title{Initial NMF fit for mixture matrix of Hi-C data}
\usage{
initial_nmf_fit(mixture, out, cell_types = 2)
}
\arguments{
\item{mixture}{Input mixture matrix for deconvolution where rows are Hi-C features and columns are samples.}

\item{out}{Complete output file path including name of output .rds}

\item{cell_types}{Integer. The number of columns in the basis matrix of the deconvolution. Corresponds to the number of cell types in bulk Hi-C mixture.}
}
\description{
Initial NMF fit for mixture matrix of Hi-C data
}
