% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout}
\alias{layout}
\title{Plot Skew-T and wind plots on a single layout}
\usage{
layout(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  convert = FALSE,
  ptop = 100,
  interpolate = TRUE,
  title = "Sample title",
  parcel = "MU",
  ...
)
}
\arguments{
\item{pressure}{- air pressure (hPa)}

\item{altitude}{- in metres}

\item{temp}{- air temperature (degree Celsius)}

\item{dpt}{- dew point temperature (degree Celsius)}

\item{wd}{- wind direction in degrees (0-360)}

\item{ws}{- wind speed in m/s}

\item{convert}{logical. Whether to convert wind speed from knots to m/s (default FALSE)}

\item{ptop}{Pressure top level to be used for plotting wind speed. Valid options should be < 200 hPa (100 by default)}

\item{interpolate}{logical, draw wind barbs only at interpolated altitudes with 1 km interval (default = TRUE)  instead of all wind barbs for a given input dataset}

\item{title}{title to be added in the header}

\item{parcel}{tracing for "MU", "ML" or "SB" parcel}

\item{...}{extra graphic arguments}
}
\description{
Jakis dluzszy opis tego co tutaj sie bedzie dzialo... Ala ma kota. ALa ma kojota
}
\examples{
data("sounding_wien")
pressure = sounding_wien$PRES
altitude = sounding_wien$HGHT
temp = sounding_wien$TEMP
dpt = sounding_wien$DWPT
wd = sounding_wien$DRCT
ws = sounding_wien$SKNT
layout(pressure, altitude, temp, dpt, wd, ws)


}
