% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_basis_matrix.R
\name{mutate_basis_matrix}
\alias{mutate_basis_matrix}
\title{Compute summary statistics on rows of cell type profile matrix}
\usage{
mutate_basis_matrix(nmf_obj)
}
\arguments{
\item{nmf_obj}{NMF object - usually the output of `initial_nmf_fit`}
}
\description{
Compute summary statistics on rows of cell type profile matrix
}
