% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northplatte.R
\docType{data}
\name{northplatte}
\alias{northplatte}
\title{Exemplary sounding dataset - sample from LBF North Platte (WMO ID: 72562) - 03 July 1999, 00:00 UTC}
\format{
A data frame with 71 rows and 6 variables as described in `get_sounding()`

\describe{
  \item{pressure}{pressure [hPa]}
  \item{altitude}{altitude [m]}
  \item{temp}{temperature [degree Celsius]}
  \item{dpt}{dew point temperature [degree Celsius]}
  \item{wd}{wind direction [azimuth as degrees]}
  \item{ws}{wind speed [knots]}
}
}
\source{
http://weather.uwyo.edu/upperair/sounding.html
}
\usage{
data("northplatte")
}
\description{
The object contains pre-downloaded sounding dataset from University of Wyoming sounding database.
Dataset can be downloaded with the following syntax: 
northplatte = get_sounding(wmo_id = 72562, yy = 1999, mm = 7, dd = 3, hh = 00)
}
\examples{
data(northplatte)
head(northplatte)
}
\keyword{datasets}
\keyword{thunder}
