% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_helpers.R
\name{parcel_helpers}
\alias{parcel_helpers}
\title{parcel_helpers}
\usage{
parcel_helpers(pressure, altitude, temp, dpt, wd, ws)
}
\arguments{
\item{pressure}{- air pressure (hPa)}

\item{altitude}{- in metres}

\item{temp}{- air temperature (degree Celsius)}

\item{dpt}{- dew point temperature (degree Celsius)}

\item{wd}{- wind direction in degrees (0-360)}

\item{ws}{- wind speed in [m/s or knots / TODO / TODISCUSS]}
}
\description{
Internal function to be used for re-calculate parcel trajectories that can be analysed or drawn
on Skew-T diagrams
}
\examples{
data("sounding_wien")
attach(sounding_wien)
sounding_plot()
output <- parcel_helpers(PRES, HGHT, TEMP, DWPT, DRCT, SKNT)
skewt_lines(output$dpt, output$pressure, col = 'forestgreen',lwd = 2.5)
skewt_lines(output$temp,output$pressure, col = 'red', lwd = 2.5)
skewt_lines(output$MU,output$pressure, col = "orange", lty = 1, lwd = 2)
skewt_lines(output$tempV,output$pressure, col = "red3", lty = 3, lwd = 1.5)
}
