% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_shading.R
\name{parcel_shading}
\alias{parcel_shading}
\title{Calculate X and Y coordinates for lines to be drawn on Skew-T diagram}
\usage{
parcel_shading(temp1, pressure1, temp2, pressure2, ptop = 100, ...)
}
\arguments{
\item{temp1}{coordinates to be used based on air temperature vector}

\item{pressure1}{coordinates to be used base on air pressure vector}

\item{temp2}{coordinates to be used based on air temperature vector}

\item{pressure2}{coordinates to be used base on air pressure vector}

\item{ptop}{upper limit of drawn trajectory (default: 100 hPa); use only if a line goes beyond the drawing area}

\item{...}{other graphical parameters that can be passed to `graphics::lines()` function, such as `lwd`, `lty`, `col`, etc.}
}
\description{
Draw any line on Skew-T diagram using temperature and pressure as coordinates
}
