% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_barbs.R
\name{sounding_barbs}
\alias{sounding_barbs}
\title{Plot wind profile using wind barbs}
\usage{
sounding_barbs(
  pres,
  ws,
  wd,
  altitude,
  convert = FALSE,
  ptop = 100,
  interpolate = TRUE,
  showaxis = FALSE,
  ...
)
}
\arguments{
\item{pres}{pressure levels}

\item{ws}{wind speed in knots(!)}

\item{wd}{wind direction in degress}

\item{altitude}{altitude in metres above sea level}

\item{convert}{logical. Whether to convert wind speed from knots to m/s (default FALSE)}

\item{ptop}{Pressure top level to be used for plotting wind speed. Valid options should be < 200 hPa (100 by default)}

\item{interpolate}{logical, draw wind barbs only at interpolated altitudes with 1 km interval (default = TRUE)  instead of all wind barbs for a given input dataset}

\item{showaxis}{logical, drawing bounding box with left axis for pressure heighs (default FALSE)}

\item{...}{extra graphic arguments}
}
\description{
Function for plotting wind direction and wind speed profile. 
TODO: Can be launched as standalone function or coupled with pre-drawn Skew-T diagram.
}
\examples{
# load examplary dataset:
data("sounding_wien")
attach(sounding_wien)
sounding_barbs(pres = PRES, ws = SKNT, wd = DRCT, altitude = HGHT, convert = TRUE)

}
