% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_export.R
\name{sounding_export}
\alias{sounding_export}
\title{Sounding export}
\usage{
sounding_export(pressure, altitude, temp, dpt, wd, ws)
}
\arguments{
\item{pressure}{- pressure [hPa]}

\item{altitude}{- altitude [meters]}

\item{temp}{- temperature [degree Celsius]}

\item{dpt}{- dew point temperature [degree Celsius]}

\item{wd}{- wind direction [azimuth in degrees]}

\item{ws}{- wind speed [kn]}
}
\description{
Internal package function for exporting interpolated profile with 5m steps
}
\examples{
data("sounding_wien")
attach(sounding_wien)
skewt_plot()
output <- sounding_export(pressure, altitude, temp, dpt, wd, ws)
skewt_lines(output$dpt, output$pressure, col = 'forestgreen',lwd = 2.5)
skewt_lines(output$temp,output$pressure, col = 'red', lwd = 2.5)
skewt_lines(output$MU,output$pressure, col = "orange", lty = 1, lwd = 2)
skewt_lines(output$tempV,output$pressure, col = "red3", lty = 3, lwd = 1.5)
}
