% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_hodograph.R
\name{sounding_hodograph}
\alias{sounding_hodograph}
\title{Plot hodograph based on rawinsonde data}
\usage{
sounding_hodograph(
  ws,
  wd,
  altitude,
  max_hght = 12000,
  max_speed = 25,
  lab_hghts = c(0, 1, 3, 6, 9, 12),
  ...
)
}
\arguments{
\item{ws}{- wind speed [kn]}

\item{wd}{- wind direction [azimuth in degrees]}

\item{altitude}{- altitudes (metres)}

\item{max_hght}{- maximum altitude to be considered on the hodograph, 12 km used by default}

\item{max_speed}{- range of the hodograph to be drawn, 25 m/s used as default}

\item{lab_hghts}{- height labels to be drawn on the hodograph, 0, 1, 3, 6, 9, 12 used by default; NULL for skipping labels}

\item{...}{other graphical parameters to be used with plot() function}
}
\description{
Plot rawinsonde hodograph to show changes of wind speed and direction in vertical profile
}
\examples{
data("sounding_wien")
attach(sounding_wien)
# plot the hodograph:
sounding_hodograph(ws, wd, altitude)
}
