% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_save.R
\name{sounding_save}
\alias{sounding_save}
\title{Save `sounding_layout` to a graphical file}
\usage{
sounding_save(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  convert = FALSE,
  ptop = 100,
  interpolate = TRUE,
  title = "",
  parcel = "MU",
  filename,
  ...
)
}
\arguments{
\item{pressure}{- air pressure (hPa)}

\item{altitude}{- in metres}

\item{temp}{- air temperature (degree Celsius)}

\item{dpt}{- dew point temperature (degree Celsius)}

\item{wd}{- wind direction in degrees (0-360)}

\item{ws}{- wind speed in m/s}

\item{convert}{logical. Whether to convert wind speed from knots to m/s (default FALSE)}

\item{ptop}{Pressure top level to be used for plotting wind speed. Valid options should be < 200 hPa (100 by default)}

\item{interpolate}{logical, draw wind barbs only at interpolated altitudes with 1 km interval (default = TRUE)  instead of all wind barbs for a given input dataset}

\item{title}{title to be added in the layout's header}

\item{parcel}{tracing for "MU", "ML" or "SB" parcel}

\item{filename}{output file name with extension indicating file format (e.g. "my_plot.png" or "my_plot.svg")}

\item{...}{other arguments that can be used with `sounding_plot` or other graphic arguments}
}
\description{
Auxiliary function to `sounding_plot` that plots a composite \
of Skew-T, hodograph and other relavant rawindsonde-derived profiles \
on a single layout and saves it into graphical file. \
Useful if graphical window is too small.
}
\examples{
data("sounding_wien")
sounding_wien = na.omit(sounding_wien)
attach(sounding_wien)
sounding_save(filename = "myfile.png", 
              PRES, HGHT, TEMP, DWPT, DRCT, SKNT, parcel = "MU", 
              title = "Wien - 2011/08/23, 12:00 UTC")


}
