% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_vienna.R
\docType{data}
\name{sounding_vienna}
\alias{sounding_vienna}
\title{Examplary sounding dataset - sample from Vienna (WMO ID: 11035) - 23 August 2011, 1200 UTC}
\format{
A data frame with 88 rows and 6 variables as described in `get_sounding()`



\describe{
  \item{pressure}{pressure [hPa]}
  \item{altitude}{altitude [m]}
  \item{temp}{temperature [degree Celsius]}
  \item{dpt}{dew point temperature [degree Celsius]}
  \item{wd}{wind direction [azimuth as degrees]}
  \item{ws}{wind speed [knots]}
}
}
\source{
http://weather.uwyo.edu/upperair/sounding.html
}
\usage{
data("sounding_vienna")
}
\description{
The object contains pre-downloaded sounding dataset from University of Wyoming sounding database.
Dataset can be downloaded with the following syntax: 
demo_dataset = get_sounding(wmo_id = 11035, yy = 2011, mm = 8, dd = 23, hh = 12)
}
\examples{
data(sounding_vienna)
head(sounding_vienna)
}
\keyword{datasets}
\keyword{thunder}
