% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_basis_matrix.R
\name{subset_init_nmf}
\alias{subset_init_nmf}
\title{Subset Initial NMF fit after Feature Selection}
\usage{
subset_init_nmf(mixture_data, nmf_obj)
}
\arguments{
\item{mixture_data}{mixture data from Hi-C experiment.}

\item{nmf_obj}{NMF object - usually the output of `initial_nmf_fit`}
}
\description{
Subset Initial NMF fit after Feature Selection
}
