% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_internet.R
\name{test_url}
\alias{test_url}
\title{Download file in a graceful way}
\usage{
test_url(link, output, quiet = FALSE)
}
\arguments{
\item{link}{character vector with URL to check}

\item{output}{character vector for output file name}

\item{quiet}{logical vector (TRUE or FALSE) to be passed to curl_download function. FALSE by default}
}
\description{
Function for downloading & testing url/internet connection according to CRAN policy
Example solution strongly based on https://community.rstudio.com/t/internet-resources-should-fail-gracefully/49199/12
as suggested by kvasilopoulos
}
\examples{
\donttest{
 link = "http://httpbin.org/status/200"
 output = tempfile()
 test_url(link = link, output = output)
}

}
