% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thunder.R
\name{thunder_estimate_CTP}
\alias{thunder_estimate_CTP}
\title{Step Two of THUNDER algorithm}
\usage{
thunder_estimate_CTP(.subset_mixture_path, n_cell_types, itter)
}
\arguments{
\item{n_cell_types}{Integer. The number of columns in the basis matrix of the deconvolution. Corresponds to the number of cell types in bulk Hi-C mixture.}

\item{itter}{Integer. Number of itterations for NMF algorithm. Default is 200.}

\item{.subset_mixture}{Matrix. Usually the output of `thunder_feature_selection`}
}
\description{
Step Two of THUNDER algorithm
}
