% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thunder.R
\name{thunder_feature_selection}
\alias{thunder_feature_selection}
\title{Step One of THUNDER algorithm.}
\usage{
thunder_feature_selection(
  path_to_mixture,
  subset_mix_out_path = NULL,
  n_cell_types,
  itter = 200,
  out_init_nmf = NULL
)
}
\arguments{
\item{path_to_mixture}{String. Path to .tsv file of bulk Hi-C samples.}

\item{n_cell_types}{Integer. The number of columns in the basis matrix of the deconvolution. Corresponds to the number of cell types in bulk Hi-C mixture.}

\item{itter}{Integer. Number of itterations for NMF algorithm. Default is 200.}

\item{out_init_nmf}{String. Default is NULL. If character, then saves the initial NMF fit as a .RDS object to specified file path.}
}
\description{
Step One of THUNDER algorithm.
}
