% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windbarbs.R
\name{windbarbs}
\alias{windbarbs}
\title{Wind barbs}
\usage{
windbarbs(cx, cy, direction, speed = NA, cex = 1)
}
\arguments{
\item{cx}{x coordinates on a plot}

\item{cy}{y coordinates on a plot}

\item{direction}{wind direction (0-360 degrees)}

\item{speed}{wind speed in m/s}

\item{cex}{symbole size. Default 1}
}
\description{
A slightly modified version of station.symbol function from the RadioSonde package.
}
