% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_plot.R
\name{sounding_plot}
\alias{sounding_plot}
\title{Plot Skew-T, hodograph and convective indices on a single layout}
\usage{
sounding_plot(
  pressure,
  altitude,
  temp,
  dpt,
  wd,
  ws,
  convert = FALSE,
  ptop = 100,
  interpolate = TRUE,
  title = "",
  parcel = "MU",
  max_speed = 25,
  hazards = FALSE,
  ...
)
}
\arguments{
\item{pressure}{- air pressure (hPa)}

\item{altitude}{- altitude in metres AMSL}

\item{temp}{- air temperature (degree Celsius)}

\item{dpt}{- dew point temperature (degree Celsius)}

\item{wd}{- wind direction in degrees (0-360)}

\item{ws}{- wind speed in m/s}

\item{convert}{logical. Whether to convert wind speed from knots to m/s (default FALSE)}

\item{ptop}{Pressure top level to be used for plotting wind speed. Valid options should be < 200 hPa (100 by default)}

\item{interpolate}{logical, draw wind barbs only at interpolated altitudes with 1 km interval (default = TRUE)  instead of all wind barbs for a given input dataset}

\item{title}{title to be added in the layout's header}

\item{parcel}{tracing for "MU", "ML" or "SB" parcel}

\item{max_speed}{max speed for defining hodograph limits}

\item{hazards}{logical. Whether to add extra information about possibility of severe weather risk (default  = FALSE)}

\item{...}{extra graphic arguments}
}
\description{
Function to plot a composite of Skew-T, hodograph and other relavant rawindsonde-derived profiles \
on a single layout
}
\examples{
data("sounding_wien")
sounding_wien = na.omit(sounding_wien)
attach(sounding_wien)
sounding_plot(pressure, altitude, temp, dpt, wd, ws, 
              parcel = "MU", title = "Wien - 2011/08/23, 12:00 UTC")

}
