% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_wind.R
\name{sounding_wind}
\alias{sounding_wind}
\title{Plot vertical wind speed profile}
\usage{
sounding_wind(pres, ws, ptop = 100, convert = FALSE, yaxs = TRUE, ...)
}
\arguments{
\item{pres}{pressure levels}

\item{ws}{wind speed in knots(!)}

\item{ptop}{Pressure top level to be used for plotting wind speed. Valid options should be < 200 hPa (100 by default)}

\item{convert}{logical. Whether to convert wind speed from knots to m/s (default FALSE)}

\item{yaxs}{logic. Whether to add labels to heights on Y lab}

\item{...}{extra graphic arguments}
}
\description{
Function for plotting wind speed profile. 
TODO: Can be launched as standalone function or coupled with pre-drawn Skew-T diagram.
}
\examples{
# load examplary dataset:
data("sounding_wien")
attach(sounding_wien)
sounding_wind(pres = PRES, ws = SKNT, convert = TRUE)
}
